+++ omega/jobs/__main__.py
@@ -3,7 +3,7 @@
 
 """
 Author: Aaron-Yang [code@jieyu.ai]
-Contributors: 
+Contributors:
 
     required coz we want the package is executable
 """
@@ -13,9 +13,8 @@ import fire
 
 from omega.jobs import start
 
+
 logger = logging.getLogger(__name__)
 
 if __name__ == "__main__":
-    fire.Fire({
-        "start": start
-    })
+    fire.Fire({"start": start})